test(library): split test_service.py per sub-service
Extract test classes from tests/services/library/test_service.py into
per-sub-service files mirroring the source layout:

- test_fetcher.py: TestCheckCancelling, TestBuildShortcutsData,
  TestFetchEnabledPlatforms, TestReconstructPlatformFromRegistry,
  TestTryIncrementalSkip, TestFullFetchPlatformRoms, TestFetchCollectionRoms.
- test_sync_orchestrator.py: TestShortcutDataFormat, TestSyncPreview,
  TestSyncApplyDelta, TestSyncCancelPreview, TestSyncControl,
  TestDoSyncErrorHandling, TestFinishSync, TestSafetyTimeoutGenerationGuard,
  TestSyncPreviewErrorHandling.
- test_reporter.py: TestReportSyncResults, TestGetSyncStats,
  TestGetRegistryPlatforms, TestGetRomBySteamAppId, TestFinalizeCoverPath,
  TestBuildRegistryEntry, TestClearSyncCache, TestReportSyncResultsCancelled.
- test_service.py: façade callable surface + cross-cutting integration
  (TestGetPlatforms, TestGetCollections, TestCollectionSyncEdgeCases, and
  the shortcut-removal tests that share the LibraryService fixture).

Shared fixtures live in conftest.py; shared helpers (_make_loop_with_executor,
_make_rom, _make_registry_entry, _page) live in _helpers.py.

FakeSettingsPersister is re-exported from the parent tests/conftest.py
via importlib so basedpyright resolves the symbol through the local
conftest instead of fighting pytest's nearest-conftest lookup.

All 169 library tests pass (2237 total); ruff, basedpyright,
import-linter, cosmic-call-bans remain clean.
